% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfs_to_hm3d}
\alias{cfs_to_hm3d}
\title{Convert Cubic Feet per Second to Cubic Hectometers per Day}
\usage{
cfs_to_hm3d(x)
}
\arguments{
\item{x}{Numeric vector of flow values in cubic feet per second.}
}
\value{
Numeric vector of flow values in cubic hectometers per day.
}
\description{
Converts volumetric flow from cubic feet per second (cfs)
to cubic hectometers per day (hm^3/day).
}
