% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmsta_hydrology.R
\name{dmsta_flow_series}
\alias{dmsta_flow_series}
\title{Run a DMSTA hydrology simulation over a time series}
\usage{
dmsta_flow_series(
  series,
  params,
  V_init = NULL,
  Qmethod = c("RK4", "Euler", "RKF45", "custom"),
  Nsteps = 4L,
  integrator_fun = NULL,
  interp_option = 2L,
  ...
)
}
\arguments{
\item{series}{Data frame of daily inputs.}

\item{params}{List of hydrologic parameters.}

\item{V_init}{Optional initial volume.}

\item{Qmethod}{Character. Hydrology integrator.}

\item{Nsteps}{Integer. Sub-steps per day.}

\item{integrator_fun}{Optional custom integrator.}

\item{interp_option}{Control-depth interpolation mode.}

\item{...}{Additional options.}
}
\value{
Object of class \code{"dmsta_result"}.
}
\description{
Simulates hydrologic behavior for a single cell over
a multi-day time series, managing initialization,
rolling diagnostics, and result aggregation.
}
\examples{
\dontrun{
# Read data (internal)
data(series)
series <- series[1:370,]; # for example, limit input file

# Data formatting
series$Qi <- cfs_to_hm3d(series$Flow) # cfs to hm3/d
series$Rain <- in_to_m(series$Rainfall) # inches to meters per day
series$Et <- in_to_m(series$ET)
series$Zcontrol <- 0/100 # meters; setting to zero to see what happens
# If you have release series; otherwise set to 0
series$Qr0 <- 0   # constrained outflow (forced Q) if used
series$Qr1 <- 0   # release 1
series$Qr2 <- 0   # release 2
series$Ci <- series$Conc

# input parameters
params <- list(
 A_cell = 2.19, # km2
 Zmin   = 2,              # cm
 Vmin   = 0,              # hm3
 Q_a = 1.0, # qcoef_a; discharge coef
 Q_b = 4.0, # qcoef_b; discharge exponent
 Zweir = 0, # cm; qcoef_offset; depth offset for outflow computation
 Q_zmin = 38,  # cm; qcoef_zmin
 Qomax = 0.0, # maximum discharge  hm3/day
 Qimax = 0, # maximum inflow  (hm3/day)
 Width = 1.55, # km
 Bypass_elev = 121.92, # z_byp; mean depth at which bypass begins (m) from input (cm)
 Seepout_Rate = 0.00789, # outflow seepage rate per unit head (m/day)/m from input cm/d/cm
 Seepout_Elev = 0.0, # elevation controling outflow seepage rate (input cm)
 Seepin_Rate  = 0.0,
 Seepin_Elev  = 0.0,
 ShutdownET = TRUE,
 force_Q_out = FALSE,
 wrap_interp = TRUE,
 Zinit = 40,  # cm; initial water column depth
 Qin_Frac = 0.22, # inflow_frac; fraction of basin flows going into this cell
 Zrelease = 0,   # cm; z_release; minimum depth for releases
 RecycleQ = 0,
 IsaNode = NULL,
 enable_P_release = FALSE,
 K_release = 0,
 IsaNode = NULL
)

 V_init <- (cm_to_m(params$Zinit) * params$A_cell)
 out <- dmsta_flow_series(V_init, series, params, Nsteps = 4)
 out$results
 }

}
